BoardConfig: Add a default TARGET_BOOTLOADER_BOARD_NAME
There is no need to break the build when a TARGET_BOOTLOADER_BOARD_NAME can not be derived from TARGET_PRODUCT.
Simply set a valid default value, and warn the user about the issue.
This is useful to guarantee compatibility with projects based on Sony's Open Devices that change TARGET_PRODUCT to fit their needs.

Change-Id: I88df4996b89552700f428fcdc5ad567b9573d22e